p3_merkle_tree

Struct MerkleTreeHidingMmcs

pub struct MerkleTreeHidingMmcs<P, PW, H, C, R, const DIGEST_ELEMS: usize, const SALT_ELEMS: usize> { /* private fields */ }

A vector commitment scheme backed by a MerkleTree.

This is similar to MerkleTreeMmcs, but each leaf is “salted” with random elements. This is done to turn the Merkle tree into a hiding commitment. See e.g. Section 3 of Interactive Oracle Proofs.

SALT_ELEMS should be set such that the product of SALT_ELEMS with the size of the value (P::Value) is at least the target security parameter.

R should be an appropriately seeded cryptographically secure pseudorandom number generator (CSPRNG). Something like ThreadRng may work, although it relies on the operating system to provide sufficient entropy.

Generics:

P: a leaf value

PW: an element of a digest

H: the leaf hasher

C: the digest compression function

R: a random number generator for blinding leaves

Implementations

impl<P, PW, H, C, R, const DIGEST_ELEMS: usize, const SALT_ELEMS: usize> MerkleTreeHidingMmcs<P, PW, H, C, R, DIGEST_ELEMS, SALT_ELEMS>

pub const fn new(hash: H, compress: C, rng: R) -> Self